requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
coverage==7.3.0
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
linecache2==1.0.0
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==5.6.0b3
traceback2==1.4.0
uritemplate==4.1.1
urllib3==2.0.4
zipp==3.16.2
